+++ packages/api-derive/src/chain/bestNumber.ts
@@ -2,10 +2,10 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { BlockNumber, Header } from '@plugnet/types/interfaces';
+import { BlockNumber } from '@plugnet/types/interfaces';
 
 import { Observable } from 'rxjs';
-import { filter, map } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 import { ApiInterfaceRx } from '@plugnet/api/types';
 
 import { drr } from '../util/drr';
@@ -24,10 +24,8 @@ import { drr } from '../util/drr';
  */
 export function bestNumber (api: ApiInterfaceRx): () => Observable<BlockNumber> {
   return (): Observable<BlockNumber> =>
-    api.rpc.chain.subscribeNewHead()
-      .pipe(
-        filter((header: Header): boolean => !!header && !!header.number),
-        map((header: Header): BlockNumber => header.number.unwrap()),
-        drr()
-      );
+    api.rpc.chain.subscribeNewHead().pipe(
+      map((header): BlockNumber => header.number.unwrap()),
+      drr()
+    );
 }
