+++ packages/api/src/promise/Api.ts
@@ -9,7 +9,7 @@ import { Observable, EMPTY } from 'rxjs';
 import { catchError, first, tap } from 'rxjs/operators';
 import { isFunction, assert } from '@plugnet/util';
 
-import ApiBase from '../Base';
+import ApiBase from '../base';
 import Combinator, { CombinatorCallback, CombinatorFunction } from './Combinator';
 
 /**
